refactor: enhance notebook readability and add detailed comments (#69)
